requirements.txt
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-mgmt-resource==22.0.0
azure-mgmt-resourcegraph==8.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
idna==3.4
isodate==0.6.1
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.24.1
oauthlib==3.2.2
pandas==1.5.3
portalocker==2.7.0
prettytable==3.6.0
pycparser==2.21
pyfiglet==0.8.post1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
termcolor==2.2.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
